scripts/openai/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o scripts/openai/requirements.txt
aiofiles==24.1.0
    # via testjoern (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
colorlog==6.9.0
    # via testjoern (pyproject.toml)
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.10.0
    # via openai
networkx==3.5
    # via testjoern (pyproject.toml)
openai==1.84.0
    # via testjoern (pyproject.toml)
pydantic==2.11.5
    # via openai
pydantic-core==2.33.2
    # via pydantic
pygraphviz==1.14
    # via testjoern (pyproject.toml)
python-dotenv==1.1.0
    # via testjoern (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via
    #   testjoern (pyproject.toml)
    #   openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
